Reviewed baseline database models
It follows closely the objects defined in lsst.verify package. The
database entities are:
- env, user, job, package, blob, measurement,
metric, spec
an their relationships are:
- 1 env : N jobs
- 1 job : N packages
- 1 job : N blobs
- 1 job : N measurements
- 1 metric : N specs
- 1 metric : N measurements

The current model is meant to support multiple execution environments,
such as jenkins or the local user environment. The metadata associated
to these environments is stored in the job table and can be used to look
up for the corresponding job object.

This implementation uses JSON() datatype in MySQL

See the https://sqr-009.lsst.io technote for details.